Module aws_sdk_autoscalingplans::model

Modules

application_source

See ApplicationSource

customized_load_metric_specification

See CustomizedLoadMetricSpecification

customized_scaling_metric_specification

See CustomizedScalingMetricSpecification

datapoint

See Datapoint

metric_dimension

See MetricDimension

predefined_load_metric_specification

See PredefinedLoadMetricSpecification

predefined_scaling_metric_specification

See PredefinedScalingMetricSpecification

scaling_instruction

See ScalingInstruction

scaling_plan

See ScalingPlan

scaling_plan_resource

See ScalingPlanResource

scaling_policy

See ScalingPolicy

tag_filter

See TagFilter

target_tracking_configuration

See TargetTrackingConfiguration

Structs

ApplicationSource

Represents an application source.

CustomizedLoadMetricSpecification

Represents a CloudWatch metric of your choosing that can be used for predictive scaling.

CustomizedScalingMetricSpecification

Represents a CloudWatch metric of your choosing that can be used for dynamic scaling as part

Datapoint

Represents a single value in the forecast data used for predictive scaling.

MetricDimension

Represents a dimension for a customized metric.

PredefinedLoadMetricSpecification

Represents a predefined metric that can be used for predictive scaling.

PredefinedScalingMetricSpecification

Represents a predefined metric that can be used for dynamic scaling as part of a target

ScalingInstruction

Describes a scaling instruction for a scalable resource in a scaling plan. Each scaling

ScalingPlan

Represents a scaling plan.

ScalingPlanResource

Represents a scalable resource.

ScalingPolicy

Represents a scaling policy.

TagFilter

Represents a tag.

TargetTrackingConfiguration

Describes a target tracking configuration to use with AWS Auto Scaling. Used with ScalingInstruction and ScalingPolicy.

Enums

ForecastDataType

LoadMetricType

MetricStatistic

PolicyType

PredictiveScalingMaxCapacityBehavior

PredictiveScalingMode

ScalableDimension

ScalingMetricType

ScalingPlanStatusCode

ScalingPolicyUpdateBehavior

ScalingStatusCode

ServiceNamespace